Move JourFerie controlle
